Backend/requirements.txt
Django==5.0.1
PyMySQL==1.1.0
djangorestframework==3.14.0
